third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/distribution-for-event-path-expected.txt

CONSOLE ERROR: line 11: Uncaught TypeError: a.createShadowRoot is not a function
